Clean up upload page
In this commit I clean up the code a little + change the functionality in such a way that the jpeg chosen is displayed in a modal that overlays the upload page.

The user can exit by tapping out of the modal or continue by pressing the continue button.

NOTE: continue has no link functionality yet as we still have to decide what the next steps are exactly